+++ src/Attempt/Defer.php
@@ -39,11 +39,29 @@ final class Defer implements Implementation
     }
 
     #[\Override]
-    public function flatMap(callable $map): Attempt
-    {
+    public function flatMap(
+        callable $map,
+        callable $exfiltrate,
+    ): self {
         $captured = $this->capture();
 
-        return Attempt::defer(static fn() => self::detonate($captured)->flatMap($map));
+        return new self(static fn() => self::detonate($captured)->flatMap($map));
+    }
+
+    #[\Override]
+    public function guard(
+        callable $map,
+        callable $exfiltrate,
+    ): self {
+        $captured = $this->capture();
+
+        return new self(static fn() => self::detonate($captured)->guard($map));
+    }
+
+    #[\Override]
+    public function guardError(): self
+    {
+        return $this;
     }
 
     #[\Override]
@@ -61,11 +79,23 @@ final class Defer implements Implementation
     }
 
     #[\Override]
-    public function recover(callable $recover): Attempt
-    {
+    public function recover(
+        callable $recover,
+        callable $exfiltrate,
+    ): self {
         $captured = $this->capture();
 
-        return Attempt::defer(static fn() => self::detonate($captured)->recover($recover));
+        return new self(static fn() => self::detonate($captured)->recover($recover));
+    }
+
+    #[\Override]
+    public function xrecover(
+        callable $recover,
+        callable $exfiltrate,
+    ): self {
+        $captured = $this->capture();
+
+        return new self(static fn() => self::detonate($captured)->xrecover($recover));
     }
 
     #[\Override]
@@ -84,21 +114,21 @@ final class Defer implements Implementation
         return Either::defer(static fn() => self::detonate($captured)->either());
     }
 
-    /**
-     * @return Attempt<R1>
-     */
     #[\Override]
-    public function memoize(): Attempt
+    public function memoize(callable $exfiltrate): Implementation
     {
-        return $this->unwrap();
+        return $exfiltrate($this->unwrap());
     }
 
     #[\Override]
-    public function eitherWay(callable $result, callable $error): Attempt
-    {
+    public function eitherWay(
+        callable $result,
+        callable $error,
+        callable $exfiltrate,
+    ): self {
         $captured = $this->capture();
 
-        return Attempt::defer(
+        return new self(
             static fn() => self::detonate($captured)->eitherWay($result, $error),
         );
     }